+++ models/serving/analytics/2-Diagnostic/analytics_customer_satisfaction_drivers.sql
@@ -13,20 +13,20 @@ SELECT
     i.issue_severity,
     
     COUNT(*) AS interaction_count,
-    ROUND(AVG(i.satisfaction_rating), 2) AS avg_satisfaction,
-    ROUND(AVG(i.duration_minutes), 1) AS avg_duration_minutes,
-    ROUND(AVG(i.sentiment_score), 3) AS avg_sentiment_score,
+    ROUND(AVG(i.satisfaction_rating)::numeric, 2) AS avg_satisfaction,
+    ROUND(AVG(i.duration_minutes)::numeric, 1) AS avg_duration_minutes,
+    ROUND(AVG(i.sentiment_score)::numeric, 3) AS avg_sentiment_score,
     
     -- Resolution metrics
     SUM(i.resolved_flag) AS resolved_count,
-    ROUND(SUM(i.resolved_flag) * 100.0 / COUNT(*), 2) AS resolution_rate_pct,
+    ROUND((SUM(i.resolved_flag) * 100.0 / COUNT(*))::numeric, 2) AS resolution_rate_pct,
     SUM(i.escalated_flag) AS escalated_count,
-    ROUND(SUM(i.escalated_flag) * 100.0 / COUNT(*), 2) AS escalation_rate_pct,
+    ROUND((SUM(i.escalated_flag) * 100.0 / COUNT(*))::numeric, 2) AS escalation_rate_pct,
     
     -- Sentiment breakdown
     SUM(i.positive_sentiment_flag) AS positive_count,
     SUM(i.negative_sentiment_flag) AS negative_count,
-    ROUND(SUM(i.negative_sentiment_flag) * 100.0 / COUNT(*), 2) AS negative_sentiment_pct,
+    ROUND((SUM(i.negative_sentiment_flag) * 100.0)::numeric / COUNT(*), 2) AS negative_sentiment_pct,
     
     -- Customer impact
     COUNT(DISTINCT i.customer_key) AS unique_customers,
