+++ models/serving/analytics/2-Diagnostic/analytics_loan_default_drivers.sql
@@ -22,11 +22,11 @@ WITH loan_performance AS (
         SUM(lp.late_payment_flag) AS late_payments,
         SUM(lp.missed_payment_flag) AS missed_payments,
         ROUND(AVG(lp.days_late), 1) AS avg_days_late,
-        ROUND(SUM(lp.late_fee), 2) AS total_late_fees,
+        ROUND(SUM(lp.late_fee)::numeric, 2) AS total_late_fees,
         
         -- Financial metrics
         AVG(lp.outstanding_balance) AS avg_outstanding_balance,
-        ROUND(SUM(lp.late_payment_flag) * 100.0 / COUNT(*), 2) AS late_payment_rate,
+        ROUND((SUM(lp.late_payment_flag) * 100.0 / COUNT(*))::numeric, 2) AS late_payment_rate,
         
         -- Risk classification
         CASE 
@@ -52,10 +52,10 @@ SELECT
     risk_level,
     
     COUNT(*) AS loan_count,
-    ROUND(AVG(late_payment_rate), 2) AS avg_late_payment_rate,
-    ROUND(AVG(avg_days_late), 1) AS avg_days_late,
-    ROUND(AVG(total_late_fees), 2) AS avg_late_fees,
-    ROUND(AVG(avg_outstanding_balance), 2) AS avg_outstanding,
+    ROUND(AVG(late_payment_rate)::numeric, 2) AS avg_late_payment_rate,
+    ROUND(AVG(avg_days_late)::numeric, 1) AS avg_days_late,
+    ROUND(AVG(total_late_fees)::numeric, 2) AS avg_late_fees,
+    ROUND(AVG(avg_outstanding_balance)::numeric, 2) AS avg_outstanding,
     
     -- Risk indicators
     SUM(CASE WHEN late_payments > 0 THEN 1 ELSE 0 END) AS accounts_with_late_payments,
