+++ models/serving/analytics/4-Inferential/analytics_customer_segment_statistical_comparison.sql
@@ -48,36 +48,36 @@ SELECT
     COUNT(*) AS sample_size,
     
     -- CLV statistics
-    ROUND(AVG(customer_lifetime_value), 2) AS mean_clv,
-    ROUND(STDDEV(customer_lifetime_value), 2) AS stddev_clv,
-    ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY customer_lifetime_value), 2) AS median_clv,
-    ROUND(PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY customer_lifetime_value), 2) AS q1_clv,
-    ROUND(PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY customer_lifetime_value), 2) AS q3_clv,
+    ROUND(AVG(customer_lifetime_value)::numeric, 2) AS mean_clv,
+    ROUND(STDDEV(customer_lifetime_value)::numeric, 2) AS stddev_clv,
+    ROUND((PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY customer_lifetime_value))::numeric, 2) AS median_clv,
+    ROUND((PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY customer_lifetime_value))::numeric, 2) AS q1_clv,
+    ROUND((PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY customer_lifetime_value))::numeric, 2) AS q3_clv,
     
     -- 95% Confidence Interval for CLV
-    ROUND(AVG(customer_lifetime_value) - 1.96 * STDDEV(customer_lifetime_value) / SQRT(COUNT(*)), 2) AS clv_ci_lower,
-    ROUND(AVG(customer_lifetime_value) + 1.96 * STDDEV(customer_lifetime_value) / SQRT(COUNT(*)), 2) AS clv_ci_upper,
+    ROUND((AVG(customer_lifetime_value) - 1.96 * STDDEV(customer_lifetime_value) / SQRT(COUNT(*)))::numeric, 2) AS clv_ci_lower,
+    ROUND((AVG(customer_lifetime_value) + 1.96 * STDDEV(customer_lifetime_value) / SQRT(COUNT(*)))::numeric, 2) AS clv_ci_upper,
     
     -- Income statistics
-    ROUND(AVG(annual_income), 2) AS mean_income,
-    ROUND(STDDEV(annual_income), 2) AS stddev_income,
-    ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY annual_income), 2) AS median_income,
+    ROUND(AVG(annual_income)::numeric, 2) AS mean_income,
+    ROUND(STDDEV(annual_income)::numeric, 2) AS stddev_income,
+    ROUND((PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY annual_income))::numeric, 2) AS median_income,
     
     -- Credit score statistics
-    ROUND(AVG(credit_score), 0) AS mean_credit_score,
-    ROUND(STDDEV(credit_score), 2) AS stddev_credit_score,
+    ROUND(AVG(credit_score)::numeric, 0) AS mean_credit_score,
+    ROUND(STDDEV(credit_score)::numeric, 2) AS stddev_credit_score,
     
     -- Transaction statistics
-    ROUND(AVG(transaction_count_90d), 2) AS mean_transactions,
-    ROUND(STDDEV(transaction_count_90d), 2) AS stddev_transactions,
+    ROUND(AVG(transaction_count_90d)::numeric, 2) AS mean_transactions,
+    ROUND(STDDEV(transaction_count_90d)::numeric, 2) AS stddev_transactions,
     
     -- Balance statistics
-    ROUND(AVG(total_balance), 2) AS mean_balance,
-    ROUND(STDDEV(total_balance), 2) AS stddev_balance,
+    ROUND(AVG(total_balance)::numeric, 2) AS mean_balance,
+    ROUND(STDDEV(total_balance)::numeric, 2) AS stddev_balance,
     
     -- Churn risk statistics
-    ROUND(AVG(churn_risk_score) * 100, 2) AS mean_churn_risk_pct,
-    ROUND(STDDEV(churn_risk_score) * 100, 2) AS stddev_churn_risk_pct,
+    ROUND((AVG(churn_risk_score) * 100)::numeric, 2) AS mean_churn_risk_pct,
+    ROUND((STDDEV(churn_risk_score) * 100)::numeric, 2) AS stddev_churn_risk_pct,
     
     CURRENT_TIMESTAMP AS last_updated
